Improve SIDP, serialization, and network data handling
 + Added the verbosity command to the console which modifies the messenger
   verbosity level
 * Moved UI handle color setting from the creator to a function in UIElement
 * Removed serialization escaping for Instance physics bodies
 + Add deserialization functions for Instances and physics bodies which
   directly take the data map
 + Added Room::transfer_instances() which properly transfers persistent
   instances between rooms
 * Made SIDP string interpretation the default for the string constructor
 * Standardized SIDP vector interpretation with vector_deserialize()
 + Added support for SIDP map interpretation
 * Added vector and map support to SIDP::to_str()
 * Added SIDP container bounds checking
 - Removed old SIDP accessors in favor of the macros
 + Added SIDP::v() and SIDP::m() and their associated macros to get vector and
   map references
 * Added proper pointer copying to the SIDP operator=
 + Added support for C strings to the SIDP operator=
 + Added more E_COMPUTATION types for later use with physics collision filtering
 + Added an id to the NetworkEvent class
 * Moved net::send_packet() and recv_packet() from bee/net/network.cpp to
   bee/net/packet.cpp
 * Separated net::handle_events() into multiple internal functions
 * Replaced std::unique_ptr() constructors with std::make_unique()
        ! I also tried using std::make_shared() in the messenger but I ran into
          argument errors
 * Fixed network data map handling
 * Changed physics body deserialization to not modify the shape if the type
   matches the previous type
 * Changed the RGBA constructor to use ints instead of Uint8s
 * Moved physics world construction from reset_properties() to init()
 * Changed physics instance map in Room::collision_internal() and
   check_collision_filter() to a const reference
   * Changed a logn() unit test to use doctest::Approx() to fix a case on 32bit
 * Fixed a memory leak in ObjUIButton due to a missed variable name change
 * Added nested container handling to the util function splitv()
 * Added scientific notation support to the util function is_str_floating()
 * Improved is_str_floating() and is_str_integer() performance by changing the
   digit list to a static const
        ! There's still probably a better way to write that functionality but
          it works much better than previously
 * Standardized network_map_encode() and _decode() with map_serialize() and
   _deserialize()
 + Added vector_serialize() and _deserialize() to easier convert vectors to
   strings and vice versa
 * Changed the map_serialize() and _deserialize() templates to allow maps with
   any type key rather than only strings
 * Changed map_deserialize() to modify its map in-place
 * Standardized print_map() with map_serialize() pretty-printing

+++ bee/render/rgba.cpp
@@ -18,7 +18,7 @@ namespace bee {
 		b(0),
 		a(0)
 	{}
-	RGBA::RGBA(Uint8 nr, Uint8 ng, Uint8 nb, Uint8 na) :
+	RGBA::RGBA(int nr, int ng, int nb, int na) :
 		r(nr),
 		g(ng),
 		b(nb),
